requirements.txt
Django==1.5
requests
python-keystoneclient
python-novaclient
djangorestframework